+++ MySongSpace_back/src/main/java/com/hyeonmusic/MySongSpace/service/FileService.java
@@ -1,7 +1,6 @@
 package com.hyeonmusic.MySongSpace.service;
 
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.model.AmazonS3Exception;
+
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.hyeonmusic.MySongSpace.common.utils.FileType;
 import com.hyeonmusic.MySongSpace.entity.FilePath;
@@ -11,17 +10,12 @@ import com.hyeonmusic.MySongSpace.exception.FileUploadException;
 import com.hyeonmusic.MySongSpace.exception.S3UploadException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
-
 import java.io.IOException;
 import java.net.URLDecoder;
 import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.UUID;
-
 import static com.hyeonmusic.MySongSpace.exception.utils.ErrorCode.*;
 import static com.hyeonmusic.MySongSpace.exception.utils.ErrorCode.FILE_UPLOAD_FAILED;
 
@@ -30,11 +24,7 @@ import static com.hyeonmusic.MySongSpace.exception.utils.ErrorCode.FILE_UPLOAD_F
 @RequiredArgsConstructor
 public class FileService {
 
-
-    private final AmazonS3 amazonS3;
-
-    @Value("${cloud.aws.s3.bucket}")
-    private String bucketName;
+    private final S3Service s3Service;
 
     //파일 여러개 업로드
     public FilePath uploadTrackFileAndTrackCoverFile(MultipartFile trackMusicFile, MultipartFile trackCoverFile) {
@@ -47,11 +37,7 @@ public class FileService {
             return filePath;
         } catch (Exception e) {
             deleteIfFileUploaded(musicPath, coverPath);
-            if (e instanceof S3UploadException) {
-                throw new S3UploadException(S3_SERVICE_ERROR);
-            } else {
-                throw new FileUploadException(FILE_UPLOAD_FAILED);
-            }
+            throw new FileUploadException(FILE_UPLOAD_FAILED);
         }
     }
 
@@ -59,26 +45,30 @@ public class FileService {
     public String uploadFile(MultipartFile file, String fileType) throws IOException {
         String fullFilePath = getFolderPath(fileType) + generateUniqueFileName(file.getOriginalFilename());
         ObjectMetadata metadata = createMetadata(file);
-        return uploadToS3(file, fullFilePath, metadata);
+        return s3Service.uploadToS3(file, fullFilePath, metadata);
     }
 
     public void deleteFile(String filePath) {
         try {
-            if (filePath.startsWith("/")) {
-                filePath = filePath.substring(1);
-            }
-            String decodedFilePath = URLDecoder.decode(filePath, StandardCharsets.UTF_8.toString());
-            log.info("삭제할 파일={}", decodedFilePath);
-            if (!amazonS3.doesObjectExist(bucketName, decodedFilePath)) {
-                log.info("삭제할 파일이 스토리지에 없음");
+            String decodedFilePath = normalizeFilePath(filePath);
+            if (!s3Service.doesFileExist(decodedFilePath)) {
                 throw new FileNotFoundException(FILE_NOT_FOUND);
             }
-            amazonS3.deleteObject(bucketName, decodedFilePath);
+            s3Service.deleteFileToS3(filePath);
         } catch (Exception e) {
-            //위의 경우를 제외한 모든 예외는 하나로 통일
-            log.error("삭제 버그 에러 메시지", e);
             throw new FileDeleteException(FILE_DELETE_FAILED, e.getMessage());
         }
+    }
+
+    //파일 정규화 작업(삭제시 S3에서 / 작업을 처리해줘야함)
+    public String normalizeFilePath(String filePath) {
+        if (filePath == null || filePath.isEmpty()) {
+            return filePath;
+        }
+        if (filePath.startsWith("/")) {
+            filePath = filePath.substring(1);
+        }
+        return URLDecoder.decode(filePath, StandardCharsets.UTF_8);
     }
 
     // 폴더 경로 결정
@@ -113,12 +103,5 @@ public class FileService {
         metadata.setContentType(file.getContentType());
         return metadata;
     }
-
-    // S3 업로드 수행
-    public String uploadToS3(MultipartFile file, String fullFilePath, ObjectMetadata metadata) throws IOException {
-        amazonS3.putObject(bucketName, fullFilePath, file.getInputStream(), metadata);
-        return amazonS3.getUrl(bucketName, fullFilePath).getPath().toString();
-    }
-
 
 }
